Feat: remove contracts to get user data and customer addresses.
Create traits to get user data and customer addresses and format customer data.

Refs: #3
