--- src/Contracts/AppendUserData.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace OutMart\Laravel\Customers\Contracts;
-
-interface AppendUserData
-{
-    public function getUserIdAttribute();
-
-    public function getNameAttribute();
-
-    public function getEmailAttribute();
-}
