--- src/Contracts/WithAddresses.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace OutMart\Laravel\Customers\Contracts;
-
-interface WithAddresses
-{
-}
