+++ src/Models/Customer.php
@@ -3,23 +3,11 @@
 namespace OutMart\Laravel\Customers\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use OutMart\Laravel\Customers\Contracts\AppendUserData;
-use OutMart\Laravel\Customers\Contracts\WithAddresses;
+// use OutMart\Laravel\Customers\Traits\WithAddresses;
+// use OutMart\Laravel\Customers\Traits\WithUserData;
 
 class Customer extends Model
 {
-    public function __construct()
-    {
-        if ($this instanceof AppendUserData) {
-            $this->with = ['customerable'];
-            $this->appends = ['user_id', 'name', 'email'];
-        }
-
-        if ($this instanceof WithAddresses) {
-            $this->with = ['customerable', 'addresses'];
-        }
-    }
-
     /**
      * The table associated with the model.
      *
@@ -44,35 +32,4 @@ class Customer extends Model
     protected $casts = [
         'metadata' => 'array',
     ];
-
-    public function getUserIdAttribute()
-    {
-        if ($this instanceof AppendUserData) {
-            return $this->customerable->id;
-        }
-    }
-
-    public function getNameAttribute()
-    {
-        if ($this instanceof AppendUserData) {
-            return $this->customerable->name;
-        }
-    }
-
-    public function getEmailAttribute()
-    {
-        if ($this instanceof AppendUserData) {
-            return $this->customerable->email;
-        }
-    }
-
-    public function customerable()
-    {
-        return $this->morphTo();
-    }
-
-    public function addresses()
-    {
-        return $this->hasMany(Address::class);
-    }
 }
